dirac kernel now configurable by scalar inputs, 2-4-8-16 pipes tested
The system is now designed such that a single dirac kernel should be synthesized and the same kernel is used for calculating dirac for different settings (ieo, isign, and whether to do a subtraction. The CPU code is cleaned and has functions to transfer the data back and forth from CPU to DFE LMEM, and apply the dirac operator. Multi-pipe code is implemented and tested. Number of pipe can be 2, 4, 8, or 16

	modified:   CPUCode/S_LQCDCpuCode.c
	modified:   CPUCode/in_spinor.txt
	modified:   CPUCode/out_spinor.txt
	modified:   EngineCode/src/s_lqcd/DFESu3.maxj
	modified:   EngineCode/src/s_lqcd/HalfSpinorBuffer.maxj
	modified:   EngineCode/src/s_lqcd/QCDConstants.maxj
	modified:   EngineCode/src/s_lqcd/S_LQCDKernel.maxj
	modified:   EngineCode/src/s_lqcd/S_LQCDManager.maxj
	modified:   RunRules/Simulation/.maxcompiler-activity.be
	modified:   RunRules/Simulation/build.log
	modified:   RunRules/Simulation/include/S_LQCD.h