separated compute and store datatypes (suitable for fixed point)
In LQCDConstants class, the data type for storing data in the LMEM and compute data on the chip are differentiated. The data is cast automatically to compute type at inputs and cast back to store type at outputs. The main reason for this was to allow fixed point arithmetic while easily using the same floats sample input on the CPU

	modified:   EngineCode/src/lqcd/DFESpinor.maxj
	modified:   EngineCode/src/lqcd/DFESu3.maxj
	modified:   EngineCode/src/lqcd/DFESu3Vector.maxj
	modified:   EngineCode/src/lqcd/Dslash.maxj
	modified:   EngineCode/src/lqcd/LQCDConstants.maxj
	modified:   EngineCode/src/lqcd/LQCDEngineParameters.maxj
	modified:   RunRules/Simulation/include/LQCD.h